[refactor] retirei códigos de navegação
