+++ frontend/src/screen/cadastro/perfil_pessoal/cadastroEmail.tsx
@@ -1,23 +1,21 @@
 import { useState } from "react";
 import { View, Text, StyleSheet, TouchableOpacity, TextInput, Image } from "react-native";
 
-function TelaCadEmail(navigation) {
+function TelaCadEmail (){
 
-    const [email, setEmail] = useState("")
-    const [password, setPassword] = useState("")
+    const [email, setEmail] = useState<string>("");
+    const [password, setPassword] = useState<string>("");
 
     return (
         <View style={styles.container}>
             <View style={styles.cima}>
-                <TouchableOpacity
-                    style={styles.volta}
-                    onPress={() => navigation.navigate('paginaAnterior')}>
+                <TouchableOpacity style={styles.volta}>
                     <Image source={require('../assets/setaVolta.png')}></Image>
                 </TouchableOpacity>
-                <Text style={{fontSize: 20}}>(2/5)</Text>
+                <Text style={{ fontSize: 20 }}>(2/5)</Text>
             </View>
 
-            <Image source={require('../assets/image02.png')} style={styles.image} resizeMode = 'contain' />
+            <Image source={require('../assets/image02.png')} style={styles.image} resizeMode='contain' />
             <Text style={styles.textgeral}>Insira seu email</Text>
             <TextInput
                 value={email}
@@ -32,14 +30,11 @@ function TelaCadEmail(navigation) {
                 style={styles.input}
             />
 
-            <TouchableOpacity
-                style={styles.button}
-                onPress={() => navigation.navigate('Pagina3')}>
-
+            <TouchableOpacity style={styles.button}>
                 <Text style={{ color: "#FFFFFF", fontSize: 30 }}>CONTINUAR</Text>
             </TouchableOpacity>
         </View>
-    )
+    );
 }
 
 export default TelaCadEmail;
